src: add is_fips_enabled application
This commit adds an application that queries OpenSSL to see if FIPS is
enabled. This is done by directly calling
EVP_default_properties_is_fips_enabled without loading any providers
explicitely.

The motivation for this is to show that an incorrect .include in the
OpenSSL configuration file will cause a system error if the file to be
included is not found. This is not considered a fatal error in OpenSSL
as far as I can tell but it leads to an issue in Node.js which is
documented in notes/rand-node-mksnapshot-issue.md.