Clean up the build process
This change more tightly integrates the build process to GNU make in an
effort to DRY the code and make it easier to add additional examples
without having to add a lot of unnecessary make boilerplate.

- Make `all` the default target ("goal" in GNU make speak).
- Add a "run" target.
- Build all of the examples which do not crash (the "dh" example is the
  outstanding item).
- Create the `_pre` targets using .i file extensions, given that that is
  the documented file extension for preprocessor generated files.
- Add -Wall to CFLAGS/CXXFLAGS to catch obvious issues with the examples.
- Make various knobs easier to configure without touching the Makefile
  by allowing uses to specify settings in config.mk (a file which
  purposely does not exist).

This will do for now. If I get an itch to convert this to cmake, I might
in the future; the benefit of cmake integration is cmake is more
extendable and integrates better into build tools like pkg-config.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>